fix(mint): keep mint-snackbar flow
